Issue 28017: **Assays: add new 'Sample Quantity' has typo in error message if negative number specified Issue 28018: **Assays: Confirmation of Assay Schedule delete message doesn't fit in popup window Issue 28019: **Assay Schedule: adding new timepoint does not immediately add new timepoint to right side of Assay Schedule form for it to be used

+++ study/webapp/study/vaccineDesign/Models.js
@@ -47,7 +47,7 @@ Ext4.define('LABKEY.VaccineDesign.Assay', {
         {name : 'SampleType', type : 'string'},
         {name : 'Source', type : 'string'},
         {name : 'TubeType', type : 'string'},
-        {name : 'SampleQuantity', type : 'double'},
+        {name : 'SampleQuantity', type : 'float'},
         {name : 'SampleUnits', type : 'string'},
         {name : 'VisitMap', defaultValue: []}
     ]
